refactor(llm): split LLM.cpp -- decode engine into LLM_decode.cpp (slice 3)
De-inlines the generation/decode engine out of LLMImpl.hpp into LLM_decode.cpp:
all 3 Run() overloads (embed/history/media) + GenerateKVCachePrefill (~1650 lines).
LLMImpl.hpp 3853->2214.

Handles overloaded Run + strips default args from the out-of-line defs (kept in the
header decls); the embed Run's #ifdef USE_AXCL/#else block moves verbatim.

Pure code-move, behavior-preserving. Verified on AX650N .51 BOTH backends:
- build_650 (non-AXCL) + build_x86 (AXCL) both compile+link clean
- qwen3-0.6b: 5-turn output + KV-content fingerprint BYTE-IDENTICAL to pre-split
  (same hashes as slices 1-2); Qwen3-VL-2B image path correct + deterministic

File-split complete: LLM.cpp 5155 -> LLMImpl.hpp 2214 + LLM.cpp 42 (wrappers) +
LLM_embed.cpp 376 + LLM_init.cpp 922 + LLM_decode.cpp 1654. No file over ~2200.
